Lucene index sharding

+++ community/lucene-index/src/main/java/org/neo4j/kernel/api/impl/index/LuceneDocumentRetrievalException.java
@@ -28,20 +28,20 @@ package org.neo4j.kernel.api.impl.index;
  */
 public class LuceneDocumentRetrievalException extends RuntimeException
 {
-    private int documentId;
+    private long documentId;
 
-    public LuceneDocumentRetrievalException( String message, int documentId, Throwable cause )
+    public LuceneDocumentRetrievalException( String message, long documentId, Throwable cause )
     {
-        this(message, cause);
+        this( message, cause );
         this.documentId = documentId;
     }
 
     public LuceneDocumentRetrievalException( String message, Throwable cause )
     {
-        super(message, cause);
+        super( message, cause );
     }
 
-    public int getDocumentId()
+    public long getDocumentId()
     {
         return documentId;
     }
